Fetch all entries when listing assigned producers (#5373)
This works around a confusing interface in Progenitor, which lists the
second argument to the `*_stream()` methods as a `limit`, but which
describes it as a page size. It currently is the former, a total limit
on all entries, so this change ensures we refresh our producer entire
producer list.