Blueprint `PlanningInput` API rework (#5404)
The impetus for this change was [@davepacheco's comment on
#5344](https://github.com/oxidecomputer/omicron/pull/5344#discussion_r1544638094).
Summarizing the changes, several of which are open to bikeshedding:

* The real meat of the PR: None of `PlanningInput`'s fields are public
anymore; all access have to go through methods, and none of the methods
allow mutation. The methods that iterate over sleds take a required
`SledFilter`; making use of this revealed at least one legit bug (prior
to this PR, the planner would attempt to put an NTP zone on a sled if it
was in the `(SledPolicy::Expunged, SledState::Active)` state and had no
NTP zone, because it was checking `SledState` but not `SledPolicy`).
* There is a separate `PlanningInputBuilder` that allows mutation; all
the unit tests go through this. For tests that want to take a
`PlanningInput` and modify it, they use the pattern
`PlanningInput::into_builder()` -> mutate via builder as desired ->
`PlanningInputBuilder::build()`.
* I used a `TypedUuid<SledKind>` in `PlanningInput{,Builder}`. I did not
change all the other sled IDs, so this diff is littered with
`TODO-cleanup`s where I'm having to convert between typed and untyped.
This is already big enough to review, so my preference is to do this in
a followup PR.
* `Policy` no longer includes the `sleds`; it's conceptually more like
"system-level policy". Currently it only includes the service IP pool
ranges and the target Nexus count.
* I broke `SledResources` up a bit, removing the sled state and policy.
The resources+policy+state are combined in `SledDetails`. (This was just
for clarity, because seeing "policy" under "resources" seems confusing,
but if I'm in the minority it'd be easy to change this back.)
* Internal and external DNS versions are now embedded in `PlanningInput`
instead of being separate arguments to `Planner::new_based_on()`. They
_are_ still separate arguments to
`BlueprintBuilder::build_initial_from_collection()`, because that
doesn't take a `PlanningInput` (but it still wants the DNS versions). I
think this is fine, especially since we plan to drop
`build_initial_from_collection()` in the next few weeks anyway.

I think this is worth doing for the
`SledFilter`-makes-it-harder-to-write-bugs alone, but hopefully it also
hits some of Dave's goals with reducing churn as we modify
`PlanningInput` in the future.