remove cargo registry patch for samael (#5421)
Fixes the following warning introduced by #5408:

```
warning: Patch `samael v0.0.14 (https://github.com/oxidecomputer/samael?branch=oxide/omicron#9e609a8f)` was not used in the crate graph.
Check that the patched package version and available features are compatible
with the dependency requirements. If the patch has a different version from
what is locked in the Cargo.lock file, run `cargo update` to use the new
version. This may also occur with an optional dependency that is not enabled.
```

#5408 bumped samael to v0.0.15, which brought in the upstream fix for
#4920, which is why we had the patch in the first place.