[nexus] fix up some uses of blueprints that were using the Debug impl (#5413)
Before we added the `display()` method, the way to get nicely formatted
blueprints was to use the Debug impl. While working on #5211, I noticed
that these callers were still using the old way to do things. Fix this up.

I tested this by making the `Debug` impl for `Blueprint` panic, then saw
which tests failed.

I also realized that we didn't have tests for omdb's blueprints. I believe at
the time it wasn't possible to write a test for this because we didn't
have an initial blueprint -- but now we do, so include a test.

(I also wanted to ensure that the actual blueprint ID was what we expected, so
I expanded the scope of `redact_variable` a bit. Introduce `ExtraRedactions`
so we can handle both fixed- and variable-length redactions, and use the
same logic for UUIDs.)