Adding OxQL table operations for limiting data (#5478)
- Resolves #5436
- Adds a `Limit` table op, for taking either the first or last `k`
points from any number of timeseries.
- Add OxQL grammar rules for parsing `first k` and `last k` into the
table op types.
- Add methods for coalescing the limit operations in a query, similar to
coalescing predicates. This lets us push filters through limits (and
vice versa), to implement predicate pushdown.
- Include limits when fetching data from ClickHouse. This uses the
`LIMIT BY` clause, which makes this extremely simple to implement in the
database itself. That also reveals an unfortunate interaction between
our table sort ordering and the `start_time` for cumulative metrics,
specifically those which are created before NTP is synced. That can lead
to incorrect results when using the `LIMIT BY` clause, since that takes
the first few results. That means data with an early timestamp, and
early (and random) start time, can come before the desired data with a
later timestamp (but also later start time). In this one case, with a
`last k` table op, we pay the cost of a CTE and an additional re-sorting
of the data based on timestamp.