Collect inventory when a new sled-agent comes online
Previously, we collected inventory when a sled was added through the
nexus external API. However, this is too early to gather the physical
disks in inventory for the sled, since those require a sled in the
`sled` table for each corresponding disk, before they can be added as
a control plane physical disk. This specific disk related change is
part of https://github.com/oxidecomputer/omicron/pull/5506 and without
a proper trigger on sled addition, we'd have to wait another 10 minutes
for inventory collection in order to see the disks show up. An alternate
workaround would be  to manually generate an "extra" blueprint, which
also triggers inventory collection, although this would also have to be
after the sled-agent registered itself via the nexus internal API and
got entered into the `sled` table.

Since sled-agent-put is idempotent, and inventory collection is somewhat
expensive, we only trigger collection if the sled actually gets updated in the
db. We do this via a check on the modified timestamp, which is somewhat hacky
but serves our purpose well.

H/T to @davepacheco for being [prescient](https://github.com/
oxidecomputer/omicron/pull/5066#pullrequestreview-1881455062).