[nix flake] Updates and maintainance (#5524)
This commit updates the Nix flake so that it builds with the current
dependencies on `mgd` and Clickhouse. The Clickhouse version we download
has changed, so the SHA256 hash in the Nix flake must be updated (this
can't be depended on automatically from the fs, currently, because we
store its hash as MD5 in the `clickhouse_version` file, rather than
SHA256). `mgd` now depends on OpenSSL, so I've changed the build inputs
for its derivation to include OpenSSL headers for `autoPatchelfHook`.

I think I might be the only user of the Nix flake, so I don't know if it
really makes sense to ask anyone else to review this. Rest assured that
it does, in fact, fix it. :)