[internal-dns] move sled and zone IDs to TypedUuid (#5518)
I wanted to understand how internal DNS works wrt the reconfigurator, so
I tried to port it over to using TypedUuid to basically get an overview
via compile errors. Here's the result.

I realized that we actually were overloading the `Host::Zone`'s ID to
mean either the sled ID (for dendrite) or the zone ID (for the `Other`
variant). I've addressed that in the PR by making the variant also carry the ID.
(This was always an expected outcome of the transition to typed UUIDs -- if we
really are storing more than one kind of UUID within a field, we should make
that clear by specifying our own enum variants.)

Probably depends on #5517 due to shared imports.