Drop the `service` table (#5287)
I still need to test this on madrid and I do _not_ want to merge it
before we cut the next release, but I believe this is ready for review.

Related changes / fallout also included in this PR:

* `omdb db services ...` subcommands are all gone. I believe this
functionality is covered by `omdb`'s inspection of blueprints instead.
* I removed the `SledResourceKind::{Dataset,Service,Reserved}` variants
that were unused. This isn't required, strictly speaking, but
`SledResourceKind::Service` was intended to reference the `service`
table, so I thought it was clearer to drop these for now (we can add
them back when we get to the point of using them).

There are two major pieces of functionality that now _require_ systems
to have a current target blueprint set:

* `DataStore::nexus_external_addresses()` now takes an explicit
`Blueprint` instead of an `Option<Blueprint>`. Its callers (silo
creation and reconfigurator DNS updates) fail if they cannot read the
current target blueprint.
* `DataStore::vpc_resolve_to_sleds()` now _implicitly_ requires a target
blueprint to be set, if and only if the VPC being queried involves
control plane services. (In practice today, that means the VPC ID is
exactly `SERVICES_VPC_ID`, although in the future we could have multiple
service-related VPCs.) I didn't want to make this method take an
explicit blueprint, because I think its most frequent callers are
specifying instance VPCs, which do not need to access the blueprint.

These two together mean that deploying this change to a system without a
target blueprint will result in (a) the inability to create silos or
update external DNS via reconfigurator and (b) services (including
Nexus) will not get the OPTE firewall rules they need to allow incoming
traffic. All newly-deployed systems have a (disabled) blueprint set as
of #5244, but we'll need to perform the necessary support operation to
bring already-deployed systems in line.

Closes #4947.