Separate `BlueprintZoneConfig` and `OmicronZoneConfig` (#5523)
This PR should make no changes to behavior and is solely a beefy
refactor (although the bulk of the diff volume is just in the openapi
spec; the actual changes are closer to +1000,-500). The motivating case
here is that `OmicronZoneConfig` (and particularly `OmicronZoneType`)
does not have all the information Reconfigurator needs to clean up
expunged zones; in particular, we need the CRDB ID of external IP
addresses to delete them, and `OmicronZoneType` only stores the raw IP
address. In this case we could clean this up by querying by IP to get
the ID and then delete the ID, but the plan has always been that these
types would diverge over time. We already passed that line when we added
`disposition`, but we didn't really separate the types at that point.

This ended up a little messier than I expected, because there are
several places that want to construct a `BlueprintZoneConfig` from an
`OmicronZoneConfig`+disposition. The "+ disposition" will grow to "+
disposition + external IP ID" soon, which may make this conversion
unwieldy enough to consider doing something else in all those places?
I'm not sure about this and plan to cross that bridge over the next day
or two.