Remove diesel OID cache priming (#5565)
See #5561 for rationale. We could add this back if we could figure out
how to correctly invalidate the cache after a schema upgrade, but there
don't seem to be any easy ways to do that. This will ensure correctness
for now and give us time to figure out a different path.