Reframe Omicron zone external IP creation in terms of `OmicronZoneExternalIp` (#5560)
This is PR 1 of N (probably 3?) working toward getting blueprints to
store an `OmicronZoneExternalIp` (which includes both the ID and the
snat cfg, if applicable) instead of only storing a raw `IpAddr`. It
adjusts the database query function(s) that create omicron zone IPs to
take an `OmicronZoneExternalIp` instead of the list of arguments they
took previously. The changes here include:

* `IncompleteExternalIp::for_service_explicit{,_snat}` have been
replaced by `IncompleteExternalIp::for_omicron_zone`. It now detects
based on the `OmicronZoneExternalIpKind` whether it's floating or SNAT.
* `DataStore::external_ip_allocate_service_explicit{,_snat}` have been
replaced by `DataStore::external_ip_allocate_omicron_zone`.
* `DataStore::external_ip_allocate_service{,_snat}` have been removed.
These allowed allocating an IP without specifying what the IP was (i.e.,
by taking a free IP from the pool, if one is available). We had no
non-test callers of these, and with the blueprint system wanting to make
planning decisions that assign IPs without committing to anything in
CRDB, I don't think we expect to have any callers of this in the future
(since we'll always want to try to allocate a specific IP).
* `IncompleteExternalIp::for_service{,_snat}` are gone; same reason as
above.
* `SourceNatConfig` now validates that the port pair it's given (upon
creation or deserialization) are aligned correctly; previously this
check was deferred to when we inserted the IP in the database, and the
check was an `assert!`. This is now a runtime error instead. This is
more of a drive-by-cleanup than strictly required, but I didn't like how
many places we could be passing around a potentially-invalid snat
config.
* Populated the Omicron zone NICs when constructing `PlanningInput`. We
were already fetching the nic rows from the DB, but forgot to actually
add them. This is also more of a drive-by-fix than strictly required,
but we'd need it soon anyway.