Allow different versions for differently signed RoT images
Currently we restrict all RoT archives for {Gimlet, PSC, Sidecar}
to be the same version. This restricts us to having staging and
production images to be the same version. What we actually want
is all archives for a particular kind _signed with the same key_
to have the same version.