[reconfigurator-planning] track more state while building zones (#5555)
In #5493 we'd like to track invariants like: a zone should not be added
and expunged in the same blueprint. In order to do that, we need to track
this state. (There are probably other ways to do it, but this is the most
explicit method and I really like that.)

This lives in a submodule because I don't want the rest of the blueprint
builder to reach into the internals here.

I split this from #5493 because it became somewhat complex in its own
right, with its own tests.