Remove ability to generate a Blueprint from an inventory collection (#5583)
We want to add information in blueprints, particularly
`BlueprintZoneConfig` and `BlueprintZoneType`, that is not present in
the sled-agent types `OmicronZoneConfig` and `OmicronZoneType`. Today on
main conversion between those types is bidirectional. As we add to
blueprints, we will continue to be able to convert a
`BlueprintZoneConfig` into an `OmicronZoneConfig`, but the opposite
direction will become more and more difficult as callers need to provide
the additional information required for blueprints.

We have enough users of the inventory -> blueprint direction that it's
quite painful to try to add to blueprints, so this PR attempts to knock
out one of the more common uses: converting an inventory collection into
a blueprint via `BlueprintBuilder::build_initial_from_collection()`.
This method is removed, and all the remaining changes are fallout from
that. Most uses of this have been replaced by either the blueprint
produced by `ExampleSystem` or the new
`BlueprintBuilder::build_empty_with_sleds()` helper for constructing an
empty blueprint. One test needed the full blueprint-from-inventory, so
that one actually gained a new use of converting OmicronZoneConfig ->
BlueprintZoneConfig. (Not ideal, but fine for now!)