Remove more `OmicronZoneConfig` to `BlueprintZoneConfig` conversions (#5584)
This builds on #5583; see it for rationale.

We remove half of the remaining callers that want to convert
`OmicronZoneConfig` into `BlueprintZoneConfig`:

* `nexus-test-utils`'s quasi-RSS: This now works in terms of
`BlueprintZoneConfig`, and converts down to `OmicronZoneConfig` when
necessary. This is the kind of transformation we'd like to make in real
RSS, too, although it will be more complex there.
* One unit test in `nexus-reconfigurator-execution`: replaced using the
`representative()` collection (which includes a collection from a live
system, but no blueprint) with the `example()` system (which produces
both a collection and a matching blueprint).
* Blueprint diffing - the types for the "before" side (which can be a
collection or a blueprint) are now wrapped in two-variant enums, with
some helpers for comparisons, instead of converting the collection to a
blueprint.

The remaining users of this conversion direction are the most difficult,
but at least we're down to three:

* `nexus-db-model` does this conversion to use some shared database
serialization logic. This will necessarily change when we add to
blueprints, since the serialization logic won't be shared anymore (or at
least not _as_ shared).
* RSS; see note above. We might be able to handle this like we did
`nexus-test-utils`; I'll see when it becomes necessary.
* The DNS test that uses data from a live system.

Even if the last two have to keep this conversion direction for now
(which will become more complex as data is added), at least it's only
two! That should be manageable even if the process gets a little messy.