Support env variables in virtual-hardware xtask (#5606)
Some of these environment variables were not carried over when
transitioning from the `create-virtual-hardware.sh` bash script to the
xtask. This prevented the a4x2 testbed softnpu setup from being
correctly setup, which resulted in bootstrap addresses not being
routable. This commit fixes that.

Once this is merged, then the corresponding
[fix](https://github.com/oxidecomputer/testbed/pull/43) in testbed
should work.