Periodically renew oximeter producer leases (#5512)
- Changes Nexus metric producer API. This used to return a 204, and now
returns a 201 along with a new lease duration for an oximeter producer.
- Reworks the producer server. The API endpoints it exposes were
previously available as public functions. The goal was to support using
those endpoints in an existing dropshot server, rather than a
special-purpose one just for metric production. These endpoints are now
private, so all producers use the built-in server. The point of this is
to centralize the logic for renewing a producer's lease periodically
from Nexus. The server now spawns a task for doing that internally, so
consumers only need to make the server and register their own metric
producer types (which they already did). One downside of this is that
the `base_route` field is no longer needed. That allowed placing the
collection endpoint anywhere in the Dropshot server's router. Since the
endpoint must now be served by the `oximeter_producer::Server`, that's
extraneous. New clients will have that value ignored, and set to the
expected one internally. We'll remove the field and its database column
in the future, once all clients are on the new version where it does not
matter.
- Updates the Nexus and Sled Agent metric production to use the new
server. They previously had the needed endpoints bolted onto their
existing Dropshot servers. The upshot of this is that neither needs to
bother with explicitly registering themselves with Nexus -- the producer
server does that internally, both initially and when renewing the lease.
- Fixup integration tests. These no longer need to explicitly ask Nexus
to register as a metric producer in each test, since that's done when a
Nexus object is created.