Enable the SFTP subsystem in the switch zone ssh server (#5634)
We have historically accessed the rack's switch zone from systems
running a version of OpenSSH from Helios 1.0, which uses the legacy
scp/rcp protocol for `scp(1)` operations. More recent versions of
OpenSSH default to using SFTP and only fall back to the legacy protocol
when the `scp -O` flag is used. SFTP is considered generally better and
we should enable the subsystem for this SSH server instance.