[omdb] Show physical disks in blueprints and inventory
This PR changes the construction and display of blueprint diffs
significantly. The need for the changes arose out of the tight coupling
between zones and sleds. When adding disks to the mix we needed to also
take them into account when determininig when a sled was modified. We also
want to display physical disks as tables along with the zone tables
under the sleds. This turned out to be somewhat trickier than necesary,
and so I changed how the tables were constructed and rendered. Hopefully
this will also make it easier to add new tables in the future.

One possibly controversial change is that I changed the way zone
modifications are rendered. They are no longer three lines long.
Currently only `disposition` is allowed to change, and so I used an
arrow from "old" to "new" versions. Additionally, I removed the (added/
modified/removed) suffixes as they seem redundant to me, make the lines
longer, and make things harder to read IMO. I'm open to discussion about
both of these changes. My guess is that eventually, we'll want to be
able to do row filtering, and I plan to also add some colored output,
also in this PR most likely.

It should be noted that the output mechanism is decoupled from the
representation of the tables, `BpSledSubtable`. This allows output in
other formats if desired in the future.

As for the inventory collection output, I added a requirement when
filtering on sled-id, that we also filter on collection-id, because
filtering on sled-id does a full table scan and requires a new index. I
can add the index instead if we feel it's important.

Fixes #5624