[oximeter] use `Cow<'static, str>` in self-stats
This commit changes Oximeter's `self_stats` module to use `Cow<'static,
str>` to represent the field value generated by the `FailureReason`
enum. This way, a heap-allocated string is only necessary for
dynamically formatted variants (the `FailureReason::Other` variant,
which owns an HTTP `StatusCode`). For the `FailureReason::Unreachable`
and `FailureReason::Deserialization` variants, we can just emit a
`&'static str` without allocating. Hopefully, this reduces resident
memory for the collector process somewhat when encountering a lot of
failures.