Clean up IP pool resolution from `NameOrId` (#5674)
Followup to #5660, specifically
https://github.com/oxidecomputer/omicron/pull/5660#discussion_r1583752697.
I didn't do it quite as suggested, basically because I think datastore
functions being aware of `NameOrId` is a bad thing as that is, to me, an
API concept that we should have resolved to something real by the time
we get to the datastore functions. This is still a little odd and a
little redundant, but it is cleaner and more consistent.

In each case (floating IP create, probe create, instance ephemeral IP
allocated) we are doing the following:

1. In the Nexus layer, turn `Option<NameOrId>` into
`Option<authz::IpPool>` (this is a DB query of course)
2. In the datastore, call the new `resolve_pool_for_allocation`, which
a. If a pool is specified, makes sure it's linked to the current silo,
otherwise 404
b. If a pool is not specified, get the default pool for the current silo
c. Either way, make sure we have `CreateChild` on the pool (we always do
if we're authenticated at all, but that could change in the future)
3. At this point we definitely have an `authz::IpPool` and can get on
with whatever allocation